QUANDO VOCÊ CHEGOU EM CASA, EU TE TRATEI NATURALMENTE E QUANDO FIZ AMOR CONTIGO A NOITE INTEIRA LENTAMENTE  @JonasKauan

+++ projetoAgiliz/src/main/java/agiliz/projetoAgiliz/dto/ColetasPorTempo.java
@@ -3,6 +3,6 @@ package agiliz.projetoAgiliz.dto;
 import java.time.LocalDateTime;
 
 public record ColetasPorTempo(
-        int quantidadeColetas,
+        long quantidadeColetas,
         LocalDateTime dataHora
 ) {}
